problem 2 solved
