+++ technical-test-airlines/src/app/config/general-data.ts
@@ -1,4 +1,5 @@
 export namespace GeneralData{
     export const ROUTE_VALID_LENGHT = 3;
-    export const DATA_URL = 'https://recruiting-api.newshore.es/api/flights/0';
+    export const DATA_URL = 'https://recruiting-api.newshore.es/api/flights/2';
+    export const ERROR_CONNECTION = 'Unable to connect to server';
 }
