+++ technical-test-airlines/src/app/modules/flights/search-flight/search-flight.component.css
@@ -1,81 +1,97 @@
 .background {
-    @apply w-screen h-screen relative filter brightness-50;
-  }
-  
-  .container {
-    @apply absolute top-1/2 -translate-y-2/4  transform flex justify-center items-center;
-  }
-  
-  .inicio-s {
-    @apply mb-8 text-white font-medium;
-  }
-  
-  .form {
-    max-width: 330px;
-    background: rgba(255, 255, 255, 0.6);
-    border-top: 2px solid rgba(255, 255, 255, 0.5);
-    border-left: 2px solid rgba(255, 255, 255, 0.5);
-    backdrop-filter: blur(5px);
-    @apply backdrop-blur-3xl rounded-lg m-auto w-full p-4 text-center;
-  }
-  .label {
-    @apply text-primary;
-  }
-  
-  .forgot {
-    @apply text-blue-600 hover:text-white duration-200 font-medium tracking-wider;
-  }
-  
-  .ancho {
-    width: 80%;
-  }
-  
-  .sign-in {
-    border: 2px solid white;
-    background: transparent;
-    text-transform: uppercase;
-    color: white;
-    padding: 10px 50px;
-    outline: none;
-    overflow: hidden;
-    position: relative;
-    @apply px-14 py-3 mt-8 mb-3 hover:border-blue-100 text-blue-600 hover:text-blue-800 font-black tracking-wider shadow-xl;
-  }
-  
-  span {
-    z-index: 20;  
-  }
-  
-  .sign-in:after {
-    content: '';
-      display: block;
-      position: absolute;
-      top: -36px;
-      left: -100px;
-      background: white;
-      width: 50px;
-      height: 125px;
-      opacity: 20%;
-      transform: rotate(-45deg);
-  }
-  
-  .sign-in:hover:after {
-    left: 120%;
-    transition: all 600ms cubic-bezier(0.3, 1, 0.2, 1);
-     -webkit-transition: all 600ms cubic-bezier(0.3, 1, 0.2, 1);
-  }
-  
-  .sign-in:disabled,
-  .sign-in[disabled]{
-    border: 1px solid #999999;
-    background-color: #cccccc;
-    color: #666666;
-    cursor: default;
-  }
-  .sign-in:disabled:hover:after,
-  .sign-in[disabled]:hover:after{
-    -webkit-transition: none !important;
-    -moz-transition: none !important;
-    -o-transition: none !important;
-    transition: none !important;
-  }
+  @apply w-screen h-screen relative filter brightness-50;
+}
+
+.content {
+  @apply bg-primary_light-100 h-full w-full flex justify-center;
+}
+
+.content-main {
+  @apply w-[90%] md:w-1/2 h-full pt-64;
+}
+
+.center-content {
+  @apply h-full w-full flex justify-center;
+}
+
+.center-content-main {
+  @apply w-[90%] md:w-1/2 h-full pt-64 flex justify-center;
+}
+
+.container {
+  @apply absolute top-1/2 -translate-y-2/4  transform flex justify-center items-center;
+}
+
+.inicio-s {
+  @apply mb-8 text-primary font-medium;
+}
+
+.form {
+  max-width: 330px;
+  background: rgba(255, 255, 255, 0.6);
+  border-top: 2px solid rgba(255, 255, 255, 0.5);
+  border-left: 2px solid rgba(255, 255, 255, 0.5);
+  backdrop-filter: blur(5px);
+  @apply backdrop-blur-3xl rounded-lg m-auto w-full p-4 text-center;
+}
+.label {
+  @apply text-primary;
+}
+
+.forgot {
+  @apply text-blue-600 hover:text-white duration-200 font-medium tracking-wider;
+}
+
+.ancho {
+  width: 80%;
+}
+
+.sign-in {
+  border: 2px solid white;
+  background: transparent;
+  text-transform: uppercase;
+  color: white;
+  padding: 10px 50px;
+  outline: none;
+  overflow: hidden;
+  position: relative;
+  @apply px-14 py-3 mt-8 mb-3 hover:border-blue-100 text-blue-600 hover:text-blue-800 font-black tracking-wider shadow-xl;
+}
+
+span {
+  z-index: 20;
+}
+
+.sign-in:after {
+  content: "";
+  display: block;
+  position: absolute;
+  top: -36px;
+  left: -100px;
+  background: white;
+  width: 50px;
+  height: 125px;
+  opacity: 20%;
+  transform: rotate(-45deg);
+}
+
+.sign-in:hover:after {
+  left: 120%;
+  transition: all 600ms cubic-bezier(0.3, 1, 0.2, 1);
+  -webkit-transition: all 600ms cubic-bezier(0.3, 1, 0.2, 1);
+}
+
+.sign-in:disabled,
+.sign-in[disabled] {
+  border: 1px solid #999999;
+  background-color: #cccccc;
+  color: #666666;
+  cursor: default;
+}
+.sign-in:disabled:hover:after,
+.sign-in[disabled]:hover:after {
+  -webkit-transition: none !important;
+  -moz-transition: none !important;
+  -o-transition: none !important;
+  transition: none !important;
+}
