+++ technical-test-airlines/src/app/modules/flights/search-flight/search-flight.component.ts
@@ -1,47 +1,30 @@
-
-
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-import { MatDialog } from '@angular/material/dialog';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { RouteModel } from '../../../models/route.model';
-import { GeneralData } from '../../../config/general-data';
 import { FlightModel } from 'src/app/models/flight.model';
-// import { InfoComponent } from '../../shared/components/modals/info/info.component';
-// import { UserCredentialsModel } from '../../../models/sesion/user-credentials.models';
-// import { MD5 } from 'crypto-js';
-// import { SecurityService } from 'src/app/services/shared/security.service';
-// import { RolModel } from '../../shared/modelos/rol.model';
-import { GetFlightService } from '../../../services/get-flight,service';
-import { TransportModel } from 'src/app/models/transport.model';
-import { JsonFlightsModel } from 'src/app/models/jsonflights.model';
-// import { LocalStorageService } from 'src/app/services/shared/local-storage.service';
+import { GeneralData } from '../../../config/general-data';
+import { GetFlightService } from '../../../services/models/get-flight.service';
+import { GenerateJourneyService } from '../../../services/models/generate-journey.service';
 
 @Component({
   selector: 'app-search-flight',
   templateUrl: './search-flight.component.html',
   styleUrls: ['./search-flight.component.css']
 })
+
 export class SearchFlightComponent {
-  transport = new TransportModel();
-  jsonflightList: JsonFlightsModel[] = [];
   flightList: FlightModel[] = [];
-  transportList: TransportModel[] = [];
-  form: FormGroup = new FormGroup({
-  });
-  hide: boolean = true;
-  checkToken: any;
+  form: FormGroup = new FormGroup({});
 
   constructor(
     private fb: FormBuilder,
-    private GetFlightService: GetFlightService
+    private GetFlightService: GetFlightService,
+    private GenerateJourneyService: GenerateJourneyService
   ) { }
 
   ngOnInit(): void {
     this.CreateForm();
-    this.GetFlightList();
-    /*console.log(this.flightList);
-    console.log(typeof this.flightList);*/
+    this.flightList = this.GetFlightService.GetFlightList();
   }
 
   CreateForm() {
@@ -72,9 +55,9 @@ export class SearchFlightComponent {
   SearchRoute() {
     if (!this.form.invalid) {
       this.RecordRouteForm();
-      this.GenerateJourneyData();
+      this.GenerateJourneyService.GenerateJourneyData();
       error: (error: any) => {
-        console.log('Error al conectar con el backend');
+        console.log(GeneralData.ERROR_CONNECTION);
       }
     }
   }
@@ -83,56 +66,5 @@ export class SearchFlightComponent {
     let route = new RouteModel();
     route.origin = this.form.controls["origin"].value.toUpperCase();
     route.destination = this.form.controls["destination"].value.toUpperCase();
-    console.log(route);
-  }
-
-  GenerateJourneyData() {
-    console.log(this.jsonflightList);
-    this.jsonflightList.forEach(jsonflight => {
-      
-      this.transport.flightCarrier = jsonflight.flightCarrier,
-      this.transport.flightNumber = jsonflight.flightNumber;
-      
-      this.transportList.push(this.transport);
-    });
-    console.log(this.jsonflightList);
-
-    for (let i = 0; i < this.transportList.length; i++) {
-      let flight = new FlightModel();
-      flight.destination = this.jsonflightList[i].departureStation,
-      flight.origin = this.jsonflightList[i].arrivalStation,
-
-      console.log(this.transport.flightCarrier);
-      console.log(this.transport.flightNumber);
-      
-      flight.transport!.flightCarrier = this.transport.flightCarrier,
-      flight.transport!.flightNumber = this.transport.flightNumber,
-      flight.price = this.jsonflightList[i].price;
-
-      this.jsonflightList.push(flight)
-    }
-  }
-
-  GetFlightList() {
-    this.GetFlightService.GetFlightList().subscribe({
-      next: (jsonflights: JsonFlightsModel[]) => {
-        this.jsonflightList = jsonflights;
-
-        let flight = new FlightModel();
-        let transport = new TransportModel();
-        jsonflights.forEach(jsonflight => {
-          
-          flight.destination = jsonflight.departureStation,
-          flight.origin = jsonflight.arrivalStation,
-          flight.price = jsonflight.price,
-          transport.flightCarrier = jsonflight.flightCarrier,
-          transport.flightNumber = jsonflight.flightNumber,
-          flight.transport = transport
-          this.flightList.push(flight);
-        });
-        console.log(this.flightList);
-        
-      }
-    });
   }
 }
